requirements.txt
black
flake8
jinja2
matplotlib
numpy
dotmap
pyyaml
requests
requests-cache
